+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - English dingbat circled sans-serif mode
 - English dingbat negative circled sans-serif mode
 - Tai Tham Hora numeral system
+- Tai Tham Tham numeral system
 ## [1.0] - 2025-10-20
 ### Added
 - Sora Sompeng numeral system